pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opensre"
version = "0.1.0"
description = "Open-source SRE agent for automated incident investigation and root cause analysis. Automatically analyzes alerts from Slack, Grafana, Datadog, and other tooling."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
  { name = "Tracer Cloud", email = "support@tracer.cloud" },
]
keywords = ["sre", "devops", "incident", "rca", "llm", "ai", "agent", "observability"]
dependencies = [
    # Core AI/Agent
    "anthropic>=0.77.1",
    "mcp>=1.20.0",
    "openai>=2.0.0",
    "langsmith>=0.6.8",
    "langgraph>=0.4.0",
    "langgraph-sdk",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    # Data validation
    "pydantic>=2.12.5,<3",
    "pydantic-settings>=2.12.0,<3",
    "kubernetes>=28.1.0",
    # HTTP and async
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    # Authentication
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    # AWS
    "boto3>=1.34.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "dotenv",
    "click>=8.1.0",
    "rich>=13.0.0",
    "questionary>=2.1.0",
    "PyYAML>=6.0",
    # OpenTelemetry
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "opentelemetry-instrumentation>=0.40b0",
    "opentelemetry-instrumentation-botocore>=0.40b0",
    "opentelemetry-instrumentation-requests>=0.40b0",
    # Analytics
    "posthog>=3.0.0",
    # Tracer
    "tracer_decorator",
]

[project.urls]
Homepage = "https://github.com/Tracer-Cloud/opensre"
Repository = "https://github.com/Tracer-Cloud/opensre"
Issues = "https://github.com/Tracer-Cloud/opensre/issues"
Changelog = "https://github.com/Tracer-Cloud/opensre/releases"

[project.scripts]
opensre = "app.cli.__main__:main"

[project.optional-dependencies]
sdg = [
    "sentence-transformers>=3.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=4.5.1",
    "types-requests",
    "types-PyYAML",
    "boto3-stubs[essential]",
]

[tool.setuptools.packages.find]
include = ["app*", "tests*"]

[tool.ruff.lint]
select = ["E", "F", "I", "S"]

[tool.ruff.lint.per-file-ignores]
# assert is idiomatic in pytest
"*_test.py" = ["S101"]
"tests/**/*_test.py" = ["S101"]
# vendored third-party pipeline code — not ours to fix
"tests/test_case_upstream_lambda/pipeline_code/**" = ["S"]
"tests/test_case_upstream_apache_flink_ecs/pipeline_code/**" = ["S"]
"tests/test_case_upstream_prefect_ecs_fargate/pipeline_code/**" = ["S"]
# subprocess / url-open / temp-file usage in infra helpers is intentional
"tests/shared/**" = ["S108", "S310", "S603", "S607"]
"tests/**/infrastructure_sdk/**" = ["S108", "S603", "S607"]
"tests/**/trigger_lambda/**" = ["S108", "S603", "S607"]
